requirements.txt
kaggle==1.6.6
google-api-python-client==2.121.0
numpy==1.26.4
pillow==10.2.0
albumentations==1.3.1
torch==2.1.2
torchvision==0.16.2
pandas==2.1.1
seaborn==0.13.0
scikit-learn==1.3.1
scikit-learn-extra==0.3.0
scipy==1.11.3